ENH: Use ShapedNeighborhoodRangeIterator in Watershed
ConstNeighborhoodIterator is a slow jack of all trades.

Use the experimental NeighborhoodRangeIterators to explore
performance.

The Range only work around a location, the iterating is performed
by a regular ImageIteratorWithIndex, and the index defines the location
for the range.

The random access operator[] is exercised for getting and setting pixels
in the neighborhood.

Current implementation is far from optimal, with a lot of repetitive and
expensive GetPixel calls.

Change-Id: I09f2d5a027fb518f94ad50b0729a524c6462215d